+++ app/Views/admin/jurusan/list-jurusan.php
@@ -12,14 +12,16 @@
           <td><?= $i; ?></td>
           <td><?= $value['jurusan']; ?></td>
           <td>
-            <a href="<?= base_url('admin/jurusan/edit/' . $value['id']); ?>" type="button" class="btn btn-primary p-2" id="<?= $value['id']; ?>">
-              <i class="material-icons">edit</i>
-              Edit
-            </a>
-            <button onclick='deleteItem("admin/jurusan/deleteJurusanPost","<?= $value["id"]; ?>","Konfirmasi untuk menghapus data");' class="btn btn-danger p-2" id="<?= $value['id']; ?>">
-              <i class="material-icons">delete_forever</i>
-              Delete
-            </button>
+            <?php if (isSuperadmin()) : ?>
+              <a href="<?= base_url('admin/jurusan/edit/' . $value['id']); ?>" type="button" class="btn btn-primary p-2" id="<?= $value['id']; ?>">
+                <i class="material-icons">edit</i>
+                Edit
+              </a>
+              <button onclick='deleteItem("admin/jurusan/deleteJurusanPost","<?= $value["id"]; ?>","Konfirmasi untuk menghapus data");' class="btn btn-danger p-2" id="<?= $value['id']; ?>">
+                <i class="material-icons">delete_forever</i>
+                Delete
+              </button>
+            <?php endif; ?>
           </td>
         </tr>
       <?php $i++;
