+++ app/Views/admin/kelas/list-kelas.php
@@ -14,14 +14,16 @@
           <td><b><?= $value['kelas']; ?></b></td>
           <td><?= $value['jurusan']; ?></td>
           <td>
-            <a href="<?= base_url('admin/kelas/edit/' .  $value['id_kelas']); ?>" type="button" class="btn btn-primary p-2" id="<?= $value['id_kelas']; ?>">
-              <i class="material-icons">edit</i>
-              Edit
-            </a>
-            <button onclick='deleteItem("admin/kelas/deleteKelasPost","<?= $value["id_kelas"]; ?>","Konfirmasi untuk menghapus data");' class="btn btn-danger p-2" id="<?= $value['id_kelas']; ?>">
-              <i class="material-icons">delete_forever</i>
-              Delete
-            </button>
+            <?php if (isSuperadmin()) : ?>
+              <a href="<?= base_url('admin/kelas/edit/' .  $value['id_kelas']); ?>" type="button" class="btn btn-primary p-2" id="<?= $value['id_kelas']; ?>">
+                <i class="material-icons">edit</i>
+                Edit
+              </a>
+              <button onclick='deleteItem("admin/kelas/deleteKelasPost","<?= $value["id_kelas"]; ?>","Konfirmasi untuk menghapus data");' class="btn btn-danger p-2" id="<?= $value['id_kelas']; ?>">
+                <i class="material-icons">delete_forever</i>
+                Delete
+              </button>
+            <?php endif; ?>
           </td>
         </tr>
       <?php $i++;
